Path Unit AI fix, Shared Vision problems
Without periodically reordering units to move/attack move to a certain location you can not 100% be sure, that the unit will reach this location, as orders might change or get canceled in certain situations.
To prevent this issue, units are spawned at the target location and then moved to their supposed spawn location, because AI units return to the place they were created, when the have no order.
Because units are spawned inside the enemy base, players must not have shared vision with the path units and thus with the army player. To still have shared vision, it is manually activated for every path unit a short time after it spawns.
Every other unit (e.g) buildings also need to share vision, so I will have to manually share their vision.
On startup all units owned by an army player are enumerated and share vision.