Item Shop initialization order
After all items are created their references are used to create their stats.
This allows forward reference in theory.
Though the id must be known, so this is not possible in the current state.

Items do not use abilities. Instead every hero has a stat system consisting of an absolute and relative stat bonus and a base stat (derived from the hero type and the level).
Items will change stat on pickup and drop.
This means item effect and item description are needed at both runtime and compiletime.

This allows me to stuff like:
increases strength by 10%
disables mana regeneration (reduce by 1000%, minimum is 0 mana regen)