Sea Druid + Damage Reduction Fix
Damage Reduction:
Previously a damage event would first reduce the damage from the previous damage event, if that had not been done already. However with simultaneous damage events, for example damage events that trigger damage events, the damage was not reduced, because hp was set before damage was taken.
Now there is always a 0 seconds timer implemented with doAfter().