Nullpointer Exception
Nullpointer exception when calling UnitWrapper.UnitWrapper_UnitWrapper_createBuffList

Buff, line 442
LinkedList, line 149
Buff, line 441
Buff, line 437
Buff, line 447

This is the first timt the bug occured. It was at the end of a 40 minute game with the Freezer.
The wurst stack suggests, that the bug is connected to a periodic buff event, however the only buff with a periodic event was Freeze, that was casted some time before the bug.
The wrapper was null for some reason.
There is now an additional safety mechanism in the function where the bug probably occured: if the wrapper is null, it is created for this unit and the unit name is printed.
Usually for every possible owner of a buffList, the wrapper should have been created automatically.
Possible scenarios:
1. A buff was applied on a unit that is not supposed to have buffs. For example a dummy unit, that has no wrapper.
2. A buff was applied on a valid target, before its wrapper was created.

By printing the unit's name these two scenarios can be distinguished.