Stats + Multiboard
Added a basic stats table.
Instead of using natives like addMana, or addGold, custom functions are used to keep track of the values for stats.
Multiboard now shows icons of heroes.
Multiboard will transform into stat table at the end of the game.
However a the stat table will be formatted better in the future.