Git history May 22 2018
A computer crash during  a git push of develpoment broke the push.
As a result 400 files in the repo were marked as changed even though I only changes a few after the last commit-
All marked changes were removed and the actual changes have been restored in development2.
The master branch contains development2. For a complete history of the single changes view develpoment.