Fix Shop Bug
Pausing the heroes at start prevents them from buying item until they move out of shop range.
Instead heroes will have no skill points at the beginning.
The first 5 seconds are used to initialize abilities and AI for the hero. Because of that a hero should not be able to learn skills before the abilities are initialized.
Hero initialization is delayed to not have all initialization at the same time, which could lead to a lag spike.